PCI: Add pci_ops.{add,remove}_bus() callbacks
These callbacks will be called on every newly created bus and when a bus
is being removed, respectively. This can be used by drivers to implement
driver-specific initialization and teardown of the bus, in addition to
the architecture-specifics implemented by the pcibios_add_bus() and the
pcibios_remove_bus() functions.

Signed-off-by: Thierry Reding <treding@nvidia.com>
---
Changes in v2:
- add pci_ops.remove_bus() callback
- always call pcibios_add_bus()