mfd: Do not derive resource count from array
Users may pass in NULL as the resource count, but MFD_ARRAY_SIZE(NULL)
will yield 4 on 32-bit or 8 on 64-bit architectures. Instead, pass the
resource count explicitly and only derive it in the higher-level macro
variants where passing NULL doesn't make sense.

Note that this causes a crash with drivers such as the max77620 driver
that is under development, where MFD_CELL_NAME() is used, which passes
NULL as the resource array.

Cc: Lee Jones <lee.jones@linaro.org>
Signed-off-by: Thierry Reding <treding@nvidia.com>
---
Note: It might be good to find a way to fail compilation if NULL is ever
passed to the higher-level macros such as MFD_CELL_BASIC(), but I'm not
aware of any mechanism to do so.