drm/bridge: Unconditionally include of_node member
Conditionally including the of_node member in the structure is error
prone and can easily lead to build errors on configurations that don't
have OF enabled. Remove the #ifdef CONFIG_OF guard and make the field
unconditionally available.

DRM bridges are primarily used on OF-enabled platforms and the
additional pointer doesn't have much of an impact on the total size of
the structure otherwise.

Signed-off-by: Thierry Reding <treding@nvidia.com>