iommu/tegra-smmu: Enable DMA/IOMMU integration
Enabling the DMA/IOMMU integration code on 32-bit ARM has surprising
side-effects. We don't usually use that integration code, because of
some special requirements primarily by display and GPU related code.
However, the option is enabled by default on multi-platform kernels,
and that has in the past introduced regressions that were not found
because of the lack of testing with this option enabled.

In drivers we want to explicitly detach from any DMA/IOMMU domains so
that these drivers have exclusive control over the IOMMU domains, but
we enable the DMA/IOMMU integration code nevertheless to make sure it
gets broader testing and thereby we can hopefully avoid regressions
in the future, or at least catch them earlier.

Signed-off-by: Thierry Reding <treding@nvidia.com>