gpu: host1x: Parameterize number of syncpoint bits
The host1x incarnations on chips prior to Tegra186 used eight bits to
represent syncpoints in various syncpoint related registers. Starting
with Tegra186, the number of syncpoints exceeds 256, so eight bits no
longer suffice to represent all of them, so the field containing the
syncoint IDs has been increased to 10 bits. Parameterize this value
to allow the code to work correctly on all generations of Tegra.

Signed-off-by: Thierry Reding <treding@nvidia.com>