arm64: tegra: Support GPIO wake events on Tegra194
In order for wake-up capable GPIOs to wake up the system from sleep,
some top-level controls need to be programmed in the PMC controller.
Make the PMC controller a wake-up parent for both GPIO controllers
so that wake events can be properly routed.

While at it, add the AON GPIO controller which provides the wake up
sources for the power key, for example.

Signed-off-by: Thierry Reding <treding@nvidia.com>