serial: tegra-tcu: Do not allow building as a module
The Tegra TCU driver relies on a function implemented in the mailbox
subsystem but is not exported. To avoid link errors in builders, don't
allow the driver to be built as a module. The proper fix is to export
the mailbox API, but until that happens, use this quick fix.

Signed-off-by: Thierry Reding <treding@nvidia.com>