soc/tegra: pmc: Fix scratch register access
Use the correct scratch register base addresses when determining whether
or not the PMC is accessible from TZ only. For Tegra210 and earlier this
is not a problem because the scratch registers are in the main aperture,
but Tegra186 and later the scratch registers have moved to a separate
aperture. No Tegra186 or later platforms currently secure access to the
PMC, so this is not an issue currently.

Signed-off-by: Thierry Reding <treding@nvidia.com>