rtc: tegra: Implement suspend clock source
The suspend clock source for Tegra210 and earlier is currently
implemented in the Tegra timer driver. However, the suspend clock source
code accesses registers that are part of the RTC hardware block, so both
can step on each others' toes. In practice this isn't an issue, but
there is no reason why the RTC driver can't implement the clock source,
so move the code over to the tegra-rtc driver.

Signed-off-by: Thierry Reding <treding@nvidia.com>
---
Changes in v2:
- use clocksource mask of 42 bits (32 bits for seconds, 10 bits for ms)
- make to_tegra_rtc() static inline