arm64: tegra: Hook up memory controller to SMMU on Tegra186
The memory controller on Tegra186 needs to map its memory clients to
stream IDs before the ARM SMMU device is probed in order to ensure that
the correct stream IDs are emitted for memory requests. Model this
dependency by adding a reference to the memory controller to the ARM
SMMU's device tree node.

Signed-off-by: Thierry Reding <treding@nvidia.com>