ata: ahci_tegra: Use nvmem API to access calibration fuses
The FUSE driver implements an nvmem device and exports the various fuses
required for calibration and similar as nvmem cells. Use the nvmem API
to access these fuses to get rid of the custom API.

Not only does this decouple the drivers, but it also provides for a more
formal interaction between the two which allows the dependency between
them to be expressed in a more canonical way.

Signed-off-by: Thierry Reding <treding@nvidia.com>