mmc: tegra: Set correct DMA mask for Tegra194
The SDHCI controller found on Tegra194 supports 40-bit addressing, just
like it's predecessor on Tegra186. The reason why this was restricted to
39 bits initially was because the memory subsystem on Tegra194 uses bit
39 to convert between different sector orderings for block-linear buffer
formats. This bit must only be set by drivers that know about these
formats, but the bit must never be used in an address for an allocated
buffer. The 39-bit DMA mask was used to prevent that.

However, this can be more accurately described in device tree using a
dma-ranges property. Doing so causes the DMA API to not allocate DMA
buffers in the range where bit 39 would be set. The SDHCI controller can
therefore set the proper DMA mask of 40 bits.

Signed-off-by: Thierry Reding <treding@nvidia.com>