arm64: tegra: Disable secondary CPUs on Tegra194
According to ePAPR, CPUs that are quiescent should have the "status"
property of the corresponding device tree node set to "disabled". Since
all but the boot CPU are in a quiescent state on boot, mark all
secondary CPUs as "disabled".

Note that this does not have an impact on SMP. Linux will still be able
to boot these secondary CPUs like before.

Signed-off-by: Thierry Reding <treding@nvidia.com>