WIP: drm/nouveau: Make BAR1 support optional
The purpose of BAR1 is primarily to make memory accesses coherent.
However, some GPUs do not have BAR1 functionality. For example, the
GV11B found on the Xavier SoC is DMA coherent and therefore doesn't
need BAR1.

Implement a variant of FIFO channels that work without a mapping of
instance memory through BAR1.

XXX ensure memory barriers are in place for writes

Signed-off-by: Thierry Reding <treding@nvidia.com>