lib/scatterlist: Add contiguous DMA chunks helpers
Add a few helpers to count the number of contiguous DMA chunks found in
an SG table. This is useful to determine whether or not a mapping can be
used by drivers whose devices need contiguous memory.

These helpers were previously used by the Tegra DRM driver but may be
more universally helpful, so move them to a more central place.

Signed-off-by: Thierry Reding <treding@nvidia.com>