clocksource: Add debugfs support
Add a top-level "clocksource" directory to debugfs. For each clocksource
registered with the system, a subdirectory will be added with attributes
that can be queried to obtain information about the clocksource. At this
point in time, only the clocksource's counter is exposed.

This provides a simple but useful interface to test if the registered
clock sources work correctly. It is not meant as a way to support any
kind of userspace timekeeping system. There are better ways to do that.

Signed-off-by: Thierry Reding <treding@nvidia.com>