dt-bindings: panel: simple: Allow ddc-i2c-bus property
Simple panels can also be connected to a DDC/CI I2C bus, so list the
ddc-i2c-bus property to make such panels validate properly. Note that
the actual validation will happen against panel-common.yaml, but this
entry in panel-simple.yaml is needed so that additionalProperties is
not going to raise an error.

Signed-off-by: Thierry Reding <treding@nvidia.com>