tty: serial: tegra-tcu: Add earlycon support
Add earlycon support to the Tegra TCU driver. This can come in two
forms: if no arguments are passed to the earlycon kernel command-line
option, the TCU will be looked up via the /chosen/stdout-path property
and the TX mailbox address will be read from the "reg" property. Note
that this is slightly fake because the "reg" property doesn't actually
contain any of the registers that control the TCU (in fact the TCU is
not controlled by any registers, it works solely through the RX and TX
mailboxes).

If additional arguments are passed, they are expected to contain the
earlycon driver name (i.e. tegra_comb_uart), followed by the physical
address of the TX mailbox that is to be used.

Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
[treding@nvidia.com: merged code into tegra-tcu driver]
Signed-off-by: Thierry Reding <treding@nvidia.com>