arm64: tegra: Add "reg" property for TCU on Tegra194
Technically the TCU doesn't have any registers but works with two
mailboxes that serve as RX and TX FIFOs. However, if used as early
console via /chosen/stdout-path, the earlycon code will want to use
the "reg" property to initialize the I/O memory base address that
will in turn be used to access the TX FIFO.

Add the "reg" property to the TCU device tree node to make earlycon
support work out of the box.

Signed-off-by: Thierry Reding <treding@nvidia.com>